+++ CHANGELOG.md
@@ -4,7 +4,13 @@ All notable theme changes should be added to this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [ unreleased ]
+## [ released ]
+
+## Changed 
+
+- Sale banners on home page and sales page
+
+## [ released ]
 
 ## [ 1.11.6 ]